+++ tests/test-middle.cpp
@@ -161,7 +161,8 @@ TEMPLATE_TEST_CASE("middle import", "", options_slim_default,
         for (osmid_t i = 1; i <= 10; ++i) {
             nds.push_back(i);
             auto const &node = buffer.add_node(fmt::format(
-                "n{} x{:.7f} y{:.7f}", i, lon - i * 0.003, lat + i * 0.001));
+                "n{} x{:.7f} y{:.7f}", i, lon - static_cast<double>(i) * 0.003,
+                lat + static_cast<double>(i) * 0.001));
             mid->node(node);
         }
         mid->after_nodes();
