+++ tests/test-node-locations.cpp
@@ -64,7 +64,8 @@ TEST_CASE("node locations in more than one block", "[NoDB]")
     }
 
     for (osmid_t id = 1; id <= max_id; ++id) {
-        nl.set(id, {id + 0.1, id + 0.2});
+        nl.set(id,
+               {static_cast<double>(id) + 0.1, static_cast<double>(id) + 0.2});
     }
 
     REQUIRE(static_cast<osmid_t>(nl.size()) == max_id);
